BREAKING CHANGE: rework error handler
