--- src/@next/index.ts
@@ -1,3 +1,0 @@
-export { suggestionFromGraphQLError } from './suggestions'
-export { forwardToSentry } from './sentry-handler'
-export * from './types'
