+++ src/index.ts
@@ -1,10 +1,3 @@
-export {
-    fromApollo,
-    fromResponse,
-    extractProperties,
-    extractGQLProperties,
-    isApolloWithProps
-} from './error'
-export { getSuggestion , suggestionFromGraphQLError } from './suggestions'
+export { suggestionFromGraphQLError } from './suggestions'
 export { forwardToSentry } from './sentry-handler'
 export * from './types'
