+++ src/types.ts
@@ -1,95 +1,44 @@
-import { ErrorResponse } from 'apollo-link-error'
-import { ApolloError } from 'apollo-client'
+import { ApolloError } from "apollo-client"
 
-export type ClientErrorTypes =
-    | 'SERVICE_OFFLINE'
-    | 'CONNECTION_FAILED'
-    | 'UNKNOWN_ERROR'
-    | 'EMAIL_NOT_CONFIRMED'
-    | 'EMAIL_ALREADY_EXISTS'
-    | 'SCHEMA_UNKNOWN_FIELD'
-    | 'CANNOT_FINISH_REQUEST_WITH_ORDER'
-
-export type BackendErrorsWithProperties = {
-    ENTITY_IS_STILL_REFERENCED: {
-        primaryKey: {
-            [index: string]: string
-        },
-        entity: string,
-        referencedBy: string
-    },
-    DUPLICATED_ENTITY: {
-        keys: {
-            [key: string] : string
-        }
-    }
-}
-
-export type BackEndErrorTypes =
-    | 'AUTHENTICATION_FAILED'
+export type KeepfyErrorCode =
     | 'FORBIDDEN'
-    | 'INVALID_SESSION'
+    | 'AUTHENTICATION_FAILED'
     | 'INVALID_SUBSCRIPTION'
+    | 'INVALID_SESSION'
+    | 'EMAIL_ALREADY_EXISTS'
+    | 'EMAIL_NOT_VERIFIED'
     | 'ENTITY_NOT_FOUND'
+    | 'DUPLICATED_ENTITY'
+    | 'ENTITY_IS_STILL_REFERENCED'
+    | 'INVALID_USER_ROLE'
+    | 'INACTIVE_USER'
+    | 'INVALID_CREDIT_CARD'
+    | 'OFFLINE_PAYMENT_SERVICE'
+    | 'PAYMENT_INTEGRATION'
+    | 'PAYMENT_DATA_INCONSISTENCY'
+    | 'BUSINESS_ERROR'
     | 'INVALID_INPUT'
-    | 'INTERNAL_SERVER_ERROR'
-    | 'EMPLOYEE_INACTIVE_ON_CURRENT_BRANCH'
-    | 'BUSINESS_ERROR' // general type, mostly ignored
-    | keyof BackendErrorsWithProperties
 
-export type AllErrorTypes = ClientErrorTypes | BackEndErrorTypes
+export interface KeepfyError extends ApolloError {
+    message: string
+    code: KeepfyErrorCode
+    isNext: boolean | undefined
+}
 
-export type MessageSuggestion = {
+export type ClientErrorCode =
+    | 'UNKNOWN_ERROR'
+    | 'SCHEMA_UNKNOWN_FIELD'
+    | 'CONNECTION_FAILED'
+    | 'SERVICE_OFFLINE'
+
+export type AllErrorTypes = KeepfyErrorCode | ClientErrorCode
+
+export interface KeepfyErrorTrusted {
     type: AllErrorTypes
-    title: string
     message: string
 }
 
-export type SuggestionsMap = {
-    [key in AllErrorTypes]: MessageSuggestion & {
-        type: key
-    }
-}
-
-export type GraphQLError = {
-    code: string
-    message: string
-    path: string[]
-    properties: unknown
-}
-
-export type ApolloErrorWithProps = Omit<ApolloError, 'graphQLErrors'> & {
-    graphQLErrors: ReadonlyArray<GraphQLError & { properties: unknown }>
-}
-
-export type ExtractMessageFromError = (
-    error: ApolloError
-) => AllErrorTypes
-
-export type ExtractFromError = (message: string | null) => AllErrorTypes
-
-export type LinkErrorResponse = ErrorResponse
-
-export type Severity =
-    | 'fatal'
-    | 'error'
-    | 'warning'
-    | 'log'
-    | 'info'
-    | 'debug'
-    | 'critical'
-
-export type SentryAdapter = {
-    captureException(
-        // eslint-disable-next-line @typescript-eslint/no-explicit-any
-        exception: any
-    ): string
-    setExtras(extras: {
-        // eslint-disable-next-line @typescript-eslint/no-explicit-any
-        [key: string]: any
-    }): void
-    captureMessage(
-        message: string,
-        level?: Severity
-    ): string
+export const defaultKeepfyErrorTrusted: KeepfyErrorTrusted = {
+    type: 'UNKNOWN_ERROR',
+    message: 'Ops! Ocorreu um erro, contate o administrador'
 }
